Create func to help group arrays of tokens into statements
I decided that grouping the tokens into arrays of statements then
parsing each array into a parse tree will be easier than trying to
it on a stream of tokens. Having access to all tokens in the
statement while parsing it means I won't have to implement complex
lookahead and lookbehind functionality.