Start work on modifying the strimmer pkg into a pre-parser
I've decided to incorparate the grouping of tokens into the strimmer
pkg as they are both quite small and perform preparsing activities.
This should create significantly more readable code in the exe pkg
as well as creating a clearer high level architecture. the 'strimmer'
pkg has been renamed 'preparser' to reflect the change in purpose
and functionality.