Create Strimmer type + separate Statement from the rest of preparser
After appling Occam's razor to the my previous suggested solution
I decided that passing []token.Token was simpler and easier than
the whole Statement struct since the tokens are all that the parser
needs. Statement will remain part of the preparser but separated
internally for maintainability. Strimmer struct was then required
so that the previous token type could be tracked from strim() to
strim() without the user passing it in each time. Some name changes
occurred to maintain readability and to match the functionilties
modified purposes.