Create inital Tree, Node, and Expr types to be used be the parser
I've decided to take another stab at a parse tree. I currently
conclude that the additional flexibility of trees during execution
out weighs the simplicity of using a simple list of instructions or
expressions; this decision is based on what I've read and how I wish
to steer the parser and executors emergence.