Start refactoring parse rule tests to use simpified and explicit trees
I noticed that it was redundant to test all but the input token for
parseToken() tests as its the only one that should change. All other
tokens in the tree are still asserted but the expected and actual
should always be the zero value; all we care is that any token in
the tree not expected to change doesn't change. I also noticed that
only the input tree, its left, and its Right need to be specified
for tests since parseToken() doesn't interrogate fields in sub-trees
except the input tree's left and right kind. I intend to update the
remaining tests follwoing this commit.