Create predicate_test.go to test predicate functions
The predicates will evolve as the parsers capabilities are slowly
extended. I value behavioural testing much more than precise unit
testing but in this case I'm expecting a lot of sutle mistakes to
be made. With --what I feel-- the intrinsic complexity and my
inexperience of the problem at hand, I expect them to help greatly.
I intend to make a separate file to handly calls to rule consequence
functions.