Create rule 6 to handle multiple target IDs when assigning
To do this I had to change the consequence function to return the
node to set as the current. I also had to modify rule 4 such that
it checks to ensure the current node has no kind otherwise the
both rule 4 and rule 6 predicates will return true for the same
input; this should be avoided as it introduces ambiguaity. Keeping
the rules in rough order helps prevent this but I'd rather not rely
on it; I'd rather every state of the tree matches 0 or 1 predicate.