Change rules 1, 2, & 4 to default placing in current node
The consequence of rules 2 & 4 now create a new trunk token rather
than creating a new left one. Technically this shouldn't make a
difference to the outcome of either consequence but semantically
it mirrors rule 6 which I find easier to remember and understand.
I think of it as the new node pushing the current one down and to
the left.