Write up predicate-consequence meta rules
During a trip to town I considered the predicate-consequence rules
in an abstract manner to see if I could get a clearer picture of
them. The ones I've come up with so far should provide some form
of boundary which I can use to guide the thinking of rules so I
can do it quicker and with less errors.