Impl parser_2 using statement divide and conquer approach
Managed to implement this parser to the point where it can parse:
1) `x <- 1`
2) `a, b, n <- 1, 2, n`
This implementation took an hour which was significantly quicker
and got further than any previous attempt. Unlike the last approach
this one works on whole statements and not on streams of tokens.
A shame to be sure but I can always attempt a more elegant
solution once I'm more familiar with the topic. For the
time being I'm focusing on getting a full parser working. This is
the first attempt whereby I've shunned any sort of guide and managed
to obtain flow. While the guide language should not matter, perhaps
I should be cautious of those written in Python. It's not a language
where one would implement a parser of this sorts and certainly not
one I would use to demonstrate related concepts; the lack of typing
and explicitness significantly hindered learning.