Rename `preparser` pkg to `evaluator`
I started building functions to parse particular patterns of tokens
within the parser pkg. An idea came to me about calling these `micro
parsers`. If I could modify the `preparser` to produce
statement structures instead of token arrays I could validate the
full grammer and identify the kind of statement before passing it
to the parser. Now the parser does not need to pattern match token
arrays. The micro parsers for each kind can be created as the
feature is added.