Create farm pkg to group tokens + removed redundant code
The farm pkg still needs some more testing. Initially intended for
the farm pkg to return a statement type but decided to separate the
concern for strimming and grouping from the statement pattern
matching; the initial Statement structure has been created in a pkg
called stat. I appear to gravitate to a pattern of using a structure
paired with a kind or type all within its own pkg, I'm calling it
the `Kind Structs` pattern. After creating some more
tests for the farm pkg I will work on a statement identifier pkg
that will identify and validate the the statements produced by the
farm. I've been considering major changes to other parts of the
planned language and I think I can reduce the number of keywords to
5: `loop`, `match`, `func`, `true`, and `false`, spells can be used
for anything else. I've still got to remember that I want key-value
pairs as an inbuilt type in the future.