Remove parser pkg + rename parser_2 pkg to parser
Updated the main program to use the new parser package. Using the
newer scanner and strimming interfaces was easier than before but
I did come to the conclusion that turning the scanner into a
structure was a poor move. The scanner does not need to be one and
the Farm structure does not need reset functionality, it can simple
be disposed of once a statement has been crafted. New tasks are as
follows:
NEXT: Modify scanning to remove the Scanner structure and let
      the package user call ScanShebang directly if they wish.
THEN: Remove the reset functionality from the Farm structure but
      still panic if tokens are added once a statement end has been
      identified.
THEN: (Possible) rename the farm package and contents to something
      more appropriate.
THEN: Remove the Scroll structure and just let the executor load
      the source code directly.
THEN: Move the parser/scan/err package to parser/err so and rename
      the error to something more general, `Perror` maybe.
THEN: Update the farm and match package to use `Perror` instead of
      standard errors as an error in source code is always associated
      to a token or series of tokens. The tokens can be used to
      populate the 'Perror'.