Rewrite value pkg + add type checks for stronger typing
I've rewritten the value package to use the Kind Structs pattern as
lots of concrete types were verbose and checking types explicitly was
becomming painful. I could forsee more pain going forward and
likely would have required reflection at some point.

By moving towards stronger typing I can hopefully reduce the number
of potential logical errors from mixing types and force written code
to be more explicit and thus more readable. I hope it will also make
debugging easier as the resultant error message will highlight the
true cause of the programmers mistake and not a symptom of it --and
when the symptoms appear far from causes, monitors tend to fly out of
windows--. The nature of the introduced typing is a simple check
before an assignment that the target variables current value has a
kind that is compatible with value being assigned. Unfortunatly it's
all done during evaluation (runtime) so a challenge for the future
is to move type checking to the parser where possible.

As expected, voiding a variable `x <- _` removes the type associated
with it allowing assignment of a different type, does this subvert
the strong typing effort? My current opinion is that the answer
doesn't matter because the goal is to create readable and errorless
code. When the programmer explicitly voids a variable they are
explicitly acknowledging that the current type association is
destroyed and any reader must view it as such. If the programmer
forgets and makes mistakes later then thats on them; at least they,
and their friends, can reread the code they can inquire about each
statement with a greater likelyhood of spotting the error timely,
`Why did you void this variable good sir?`. If this turns out to cause
problems later I can always modify voiding so variables maintain
their type.

I pondered about situations such as `x, x <- 1, 2` and `x, x, <- _, 1`.
Both function as expected, both assignments are performed but unlike Go
the result of both expressions are assigned to the variable. I
experimented to see how Go handles these cases. Considering the examples
in Go
	x := 0
	f := func() int {
		x++
		return x
	}
	a, b = f(), f()
and
	x = 0
	x, x = 1, x+1
expressions are evaluated left to right as expected but no assignments
take place until all results are gathered. This means 'x' will be '1'
in the second Go example due to the initial value of 'x' being '0' in
both expressions. I wondered if I should do the same here. For the time
being I will not, it becomes possible to perform `x, x <- _, "abc"`
which is actually useful --if you consider voiding of variables an
appropriate feature--. I do worry that I can then introduce confusing
and hard to read statements such as:
	x <- 0
	x, x <- x + 2, x + 2
which will result in 'x' being '4'. I've reasoned that only an idiot
would write such eye sores; I need only ensure I am not one of them.